requirements.txt
groq
python-dotenv
bs4
requests
duckduckgo-search>=5.0.0
lxml
termcolor
readability-lxml

